daemon/hwcpipe2/device/gen/ioctl_idlc/requirements.txt
pystache>=0.6.0
PyYAML>=6.0
types-PyYAML>=6.0.7
parameterized>=0.8.1
